requirements.txt
mlflow
dagshub
dvc
bentoml
python-dotenv
ipykernel
tensorflow
pandas 
dvc
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
retrying
keras_tuner
